Float to BigDecimal in progress.

+++ src/main/java/com/callippus/water/erp/repository/TariffMasterCustomRepository.java
@@ -4,6 +4,7 @@ import com.callippus.water.erp.domain.TariffMaster;
 
 import org.springframework.data.jpa.repository.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -11,7 +12,7 @@ import java.util.List;
  * Spring Data JPA repository for the TariffMaster entity.
  */
 public interface TariffMasterCustomRepository extends JpaRepository<TariffMaster,Long> {	
-	List<java.util.Map<String, Object>> findTariffs(String can, LocalDate validFrom, LocalDate validTo, float avgKL, int unMeteredFlag, int newMeterFlag);
-	List<java.util.Map<String, Object>> getTariffs(String can, LocalDate validFrom, LocalDate validTo, float avgKL,
+	List<java.util.Map<String, Object>> findTariffs(String can, LocalDate validFrom, LocalDate validTo, BigDecimal avgKL, int unMeteredFlag, int newMeterFlag);
+	List<java.util.Map<String, Object>> getTariffs(String can, LocalDate validFrom, LocalDate validTo, BigDecimal avgKL,
 			int unMeteredFlag, int newMeterFlag);
 }
